+++ src/soloud/src/audiosource/wav/soloud_wavstream.cpp
@@ -260,7 +260,7 @@ namespace SoLoud
 					{
 						for (k = 0; k < mChannels; k++)
 						{
-							aBuffer[k * aSamplesToRead + i + j] = tmp[j * mCodec.mFlac->channels + k];
+							aBuffer[k * aBufferSize + i + j] = tmp[j * mCodec.mFlac->channels + k];
 						}
 					}
 				}
@@ -281,7 +281,7 @@ namespace SoLoud
 					{
 						for (k = 0; k < mChannels; k++)
 						{
-							aBuffer[k * aSamplesToRead + i + j] = tmp[j * mCodec.mMp3->channels + k];
+							aBuffer[k * aBufferSize + i + j] = tmp[j * mCodec.mMp3->channels + k];
 						}
 					}
 				}
@@ -330,7 +330,7 @@ namespace SoLoud
 					{
 						for (k = 0; k < mChannels; k++)
 						{
-							aBuffer[k * aSamplesToRead + i + j] = tmp[j * mCodec.mWav->channels + k];
+							aBuffer[k * aBufferSize + i + j] = tmp[j * mCodec.mWav->channels + k];
 						}
 					}
 				}